nothing

+++ CoreBluetoothiOSPeripheral/CoreBluetoothiOSPeripheral/LXCBPeripheralServer.m
@@ -113,7 +113,7 @@
 
 - (void)sendToSubscribers:(NSData *)data {
   if (self.peripheral.state != CBPeripheralManagerStatePoweredOn) {
-    LXCBLog(@"sendToSubscribers: peripheral not ready for sending state: %d", self.peripheral.state);
+    LXCBLog(@"sendToSubscribers: peripheral not ready for sending state: %ld", (long)self.peripheral.state);
     return;
   }
